+++ sdks/python/apache_beam/typehints/typed_pipeline_test.py
@@ -16,6 +16,8 @@
 #
 
 """Unit tests for the type-hint objects and decorators."""
+from builtins import str
+from past.builtins import basestring
 import inspect
 import typing
 import unittest
@@ -60,7 +62,7 @@ class MainInputTest(unittest.TestCase):
       [1, 2, 3] | beam.Map(str.upper)
 
   def test_loose_bounds(self):
-    @typehints.with_input_types(typehints.Union[int, float, long])
+    @typehints.with_input_types(typehints.Union[int, float, int])
     @typehints.with_output_types(basestring)
     def format_number(x):
       return '%g' % x
